combine some middleware as static methods of a class; rename another one
